+++ build.gradle
@@ -15,6 +15,8 @@ buildscript {
         log4j_version = '2.9.1'
         confidential_id_release_group = "com.r3.corda.lib.ci"
         confidential_id_release_version = "1.0-RC02"
+        spring_boot_version = '2.0.2.RELEASE'
+        spring_boot_gradle_plugin_version = '2.0.2.RELEASE'
     }
 
     repositories {
