+++ examples/cordapp-sweepstake/states-and-flows/src/main/kotlin/com/r3/corda/lib/accounts/examples/sweepstake/clients/WorldCupApplication.kt
@@ -1,0 +1,34 @@
+package com.r3.corda.lib.accounts.examples.sweepstake.clients
+
+import org.springframework.boot.CommandLineRunner
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.context.ApplicationContext
+import org.springframework.context.annotation.Bean
+import java.util.*
+
+@SpringBootApplication
+open class WorldCupApplication {
+
+    @Bean
+    open fun commandLineRunner(ctx: ApplicationContext): CommandLineRunner {
+        return CommandLineRunner {
+            println("Let's inspect the beans provided by Spring Boot:")
+            val beanNames = ctx.beanDefinitionNames
+            Arrays.sort(beanNames)
+            for (beanName in beanNames) {
+                println(beanName)
+            }
+            println("*******************************")
+            println("THE SERVER IS READY")
+        }
+    }
+
+    companion object {
+        @JvmStatic
+        fun main(args: Array<String>) {
+            SpringApplication.run(WorldCupApplication::class.java, *args)
+        }
+    }
+
+}
