+++ examples/cordapp-sweepstake/states-and-flows/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+server.port=8080
+config.rpc.username=user1
+config.rpc.password=test
+config.rpc.port=10005
+config.rpc.host=localhost
