+++ examples/cordapp-sweepstake/states-and-flows/src/main/resources/static/sweepstake.css
@@ -1,0 +1,89 @@
+table {
+    border-collapse: collapse;
+    width: 60%;
+}
+
+th, td {
+    padding: 5px;
+    text-align: left;
+    border-bottom: 1px solid #ddd;
+}
+
+body {
+    font: normal 16px Verdana, Arial, sans-serif;
+}
+
+div.b {
+    line-height: 1.6;
+}
+
+.button {
+    background-color:#ffffff;
+    -moz-border-radius:9px;
+    -webkit-border-radius:9px;
+    border-radius:9px;
+    border:1px solid #949494;
+    display:inline-block;
+    cursor:pointer;
+    color:#080808;
+    font-family:Arial;
+    font-size:12px;
+    padding:9px 26px;
+    text-decoration:none;
+}
+.button:hover {
+    background-color:#dbdbdb;
+}
+.button:disabled,
+.button[disabled]{
+    border: 1px solid #999999;
+    background-color: #cccccc;
+    color: #666666;
+}
+
+
+
+
+.firstCol {
+    width: 120px;
+    text-align: left;
+}
+.secondCol {
+    width: 250px;
+    text-align: left;
+}
+
+.thirdCol {
+    width: 500px;
+    text-align: left;
+}
+
+.fourthCol {
+    width: 120px;
+    text-align: left;
+}
+
+.grouping tr:nth-child(n){
+    background: #c5d5c5;
+}
+.grouping tr:nth-child(n+5){
+    background: #9fa9a3;
+}
+.grouping tr:nth-child(n+9){
+    background: #c0ded9;
+}
+.grouping tr:nth-child(n+13){
+    background: #b0aac0;
+}
+.grouping tr:nth-child(n+17){
+    background: #77a8a8;
+}
+.grouping tr:nth-child(n+21){
+    background: #e4d1d1;
+}
+.grouping tr:nth-child(n+25){
+    background: #b9b0b0;
+}
+.grouping tr:nth-child(n+29){
+    background: #d9ecd0;
+}
