+++ examples/cordapp-sweepstake/states-and-flows/src/main/resources/worldcupteams.txt
